Complete getElementsByClassName function
Create an ampty array that will be used to hold all of the elements with the className that is passed in. Create a variable that represents that body of the document. Create an inner function that traverses the DOM. Pass the entire document into the inner function. Check to see if the element has a classList property and whether or not it containts  the className that we're searching for. If it does, push the element into the empty array. Check if the element has children and if it does, loop through the children and recursively run the innerfunction on each child. Return the array full of elements with the className.